Move unit test for check_column to utils
